# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kayak rental platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/inventory:
    get:
      operationId: adminListInventory
      tags: [admin]
      summary: Admin — list inventory units
      parameters:
        - in: query
          name: status
          schema:
            type: ["string", "null"]
        - in: query
          name: product_id
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryUnit"
  /admin/inventory/{id}:
    patch:
      operationId: adminUpdateInventory
      tags: [admin]
      summary: Admin — update inventory unit status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryUnit"
  /admin/customers:
    get:
      operationId: adminListCustomers
      tags: [admin]
      summary: Admin — list customers
      parameters:
        - in: query
          name: search
          schema:
            type: ["string", "null"]
        - in: query
          name: page
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCustomersResponse"
  /admin/customers/{id}:
    get:
      operationId: adminGetCustomer
      tags: [admin]
      summary: Admin — get customer detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer with order history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCustomerDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/chat-sessions:
    get:
      operationId: adminListChatSessions
      tags: [admin]
      summary: Admin — list chat sessions
      parameters:
        - in: query
          name: status
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Chat sessions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminChatSession"
  /admin/chat-sessions/{id}/reply:
    post:
      operationId: adminReplyChatSession
      tags: [admin]
      summary: Admin — send reply in chat session
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Reply sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    AdminOrder:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        deliveryType:
          type: string
        exactPrice:
          type: ["number", "null"]
        approximatePrice:
          type: ["number", "null"]
        managerReviewRequired:
          type: boolean
        createdAt:
          type: string
        itemCount:
          type: integer
      required:
        - id
        - orderNumber
        - status
        - startDate
        - endDate
        - customerName
        - customerPhone
        - deliveryType
        - managerReviewRequired
        - createdAt
        - itemCount
...
    InventoryUnit:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        serialNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, busy, reserved, checked, ready_for_issue, in_repair, incomplete, incoming, written_off]
        condition:
          type: ["string", "null"]
        warehouseLocation:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - productId
        - productName
        - status
        - createdAt
    UpdateInventoryRequest:
      type: object
      properties:
        status:
          type: string
          enum: [available, busy, reserved, checked, ready_for_issue, in_repair, incomplete, incoming, written_off]
        condition:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        warehouseLocation:
          type: ["string", "null"]
      required:
        - status
    AdminCustomersResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminCustomer"
        total:
          type: integer
        page:
          type: integer
        perPage:
          type: integer
      required:
        - data
        - total
        - page
        - perPage
    AdminCustomer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        communicationChannel:
          type: ["string", "null"]
        orderCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - orderCount
        - createdAt
    AdminCustomerDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        communicationChannel:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        orders:
          type: array
          items:
            $ref: "#/components/schemas/AdminOrder"
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - orders
        - createdAt
...
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        sender:
          type: string
          enum: [customer, manager, bot]
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - sessionId
        - sender
        - content
        - createdAt
    AdminChatSession:
      type: object
      properties:
        id:
          type: integer
        sessionToken:
          type: string
        status:
          type: string
        customerName:
          type: ["string", "null"]
        lastMessage:
          type: ["string", "null"]
        unreadCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - sessionToken
        - status
        - unreadCount
        - createdAt
...
    SendMessageRequest:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...