# lib/api-zod/src/generated/types/adminCustomer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kayak rental platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminCustomer {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  communicationChannel?: string | null;
  orderCount: number;
  createdAt: string;
}